Instantaneous frequencies, damping ratios

FoliationsManifoldsAutoencoders.MAPFrequencyDamping — Function

That, Rhat = MAPFrequencyDamping(MW::DensePolyManifold, XW, MS::DensePolyManifold, XS, amp_max::Number; output = I)

Creates a polar representation of the system, written as

\[\begin{pmatrix}r\\ \theta \end{pmatrix} \mapsto \begin{pmatrix}r \hat{R}(r)\\ \theta + \hat{T}(r) \end{pmatrix}\]

The independent variable $r$ represent the root mean square amplitude of the vibration over one period. Also, the frequency represented by $\hat{T}(r)$ is correct with respect to the Euclidean frame.

The input parameters are

MW, XW is $\boldsymbol{W}$: decoder or manifold immersion

MS, XS is $\boldsymbol{S}$: two-dimenional nonlinear map

amp_max: maximum amplitude accurately represented by

output: an optional linear map, that transforms the output of the immersion $\boldsymbol{W}$. By default, the identity map is used.

It is assumed the the deconder and the nonlinear map satisfy the invariance equation

\[\boldsymbol{W} \circ \boldsymbol{S} = \boldsymbol{F} \circ \boldsymbol{W},\]

where $\boldsymbol{F}$ represent a dynamical system $\dot{\boldsymbol{x}} = \boldsymbol{F}(\boldsymbol{x})$.

The two-dimensional system $\boldsymbol{S}$ has a focus-type equilibrium at the origin, either in the real normal form

\[\begin{pmatrix}z_{1}\\ z_{2} \end{pmatrix} \mapsto \begin{pmatrix}z_{1}f_{r}\left(z_{1}^{2}+z_{2}^{2}\right)-z_{2}f_{r}\left(z_{1}^{2}+z_{2}^{2}\right)\\ z_{1}f_{i}\left(z_{1}^{2}+z_{2}^{2}\right)+z_{2}f_{r}\left(z_{1}^{2}+z_{2}^{2}\right) \end{pmatrix}.\]

or in the complex normal form

\[\begin{pmatrix}z\\ \overline{z} \end{pmatrix} \mapsto \begin{pmatrix} z + z f(\vert z \vert^2) \\ \overline{z} + \overline{z} \overline{f}(\vert z \vert^2) \end{pmatrix}\]

The output That, Rhat represent $\hat{T}, \hat{R}: [0,r_1) \to \mathbb{R}$, which satisfy the invariance equation

\[\hat{\boldsymbol{W}}\left(r \hat{R}(r), \theta + \hat{T}(r)\right) = \boldsymbol{F}( \hat{\boldsymbol{W}}\left(r, \theta\right) ).\]

FoliationsManifoldsAutoencoders.ODEFrequencyDamping — Function

That, Rhat = ODEFrequencyDamping(MW::DensePolyManifold, XW, MS::DensePolyManifold, XS, amp_max::Number; output = I)

Creates a polar representation of the system, written as

\[\begin{pmatrix}\dot{r}\\ \dot{\theta} \end{pmatrix} = \begin{pmatrix}r \hat{R}(r)\\ \hat{T}(r) \end{pmatrix}\]

The independent variable $r$ represent the root mean square amplitude of the vibration over one period. Also, the frequency represented by $\hat{T}(r)$ is correct with respect to the Euclidean frame.

The input parameters are

MW, XW is $\boldsymbol{W}$: decoder or manifold immersion

MS, XS is $\boldsymbol{S}$: two-dimenional nonlinear map

amp_max: maximum amplitude accurately represented by

output: an optional linear map, that transforms the output of the immersion $\boldsymbol{W}$. By default, the identity map is used.

It is assumed that the deconder and the nonlinear map satisfy the invariance equation

\[\boldsymbol{W} \cdot D\boldsymbol{S} = \boldsymbol{F} \circ \boldsymbol{W},\]

where $\boldsymbol{F}$ represent the right-hand side of a differential equation $\dot{\boldsymbol{x}} = \boldsymbol{F}(\boldsymbol{x})$.

The two-dimensional system $\boldsymbol{S}$ has a focus-type equilibrium at the origin, either in the real normal form

\[\begin{pmatrix}\dot{z}_{1}\\ \dot{z}_{2} \end{pmatrix} = \begin{pmatrix}z_{1}f_{r}\left(z_{1}^{2}+z_{2}^{2}\right)-z_{2}f_{r}\left(z_{1}^{2}+z_{2}^{2}\right)\\ z_{1}f_{i}\left(z_{1}^{2}+z_{2}^{2}\right)+z_{2}f_{r}\left(z_{1}^{2}+z_{2}^{2}\right) \end{pmatrix}.\]

or in the complex normal form

\[\begin{pmatrix}\dot{z}\\ \dot{\overline{z}} \end{pmatrix} = \begin{pmatrix} z + z f(\vert z \vert^2) \\ \overline{z} + \overline{z} \overline{f}(\vert z \vert^2) \end{pmatrix}\]

The output That, Rhat represent $\hat{T}, \hat{R}: [0,r_1) \to \mathbb{R}$, which satisfy the invariance equation

\[D_1\hat{\boldsymbol{W}}\left(r, \theta\right)r \hat{R}(r) + D_2\hat{\boldsymbol{W}}\left(r, \theta\right)\hat{T}(r) = \boldsymbol{F}( \hat{\boldsymbol{W}}\left(r, \theta\right) ).\]