Polynomials

FoliationsManifoldsAutoencoders.DensePolyManifold — Type

M = DensePolyManifold(ndim, n, order; min_order = 0, identity = false, field::AbstractNumbers=ℝ)

Creates a manifold structure of a dense polynomial with ndim input variables and n output dimension of maximum order order. One can set the smallest represented order by min_order. The parameter field can be set to either real ℝ or complex ℂ.

Base.zero — Method

X = zero(M::DensePolyManifold)

Create a representation of a zero polynomial with manifold structure M.

FoliationsManifoldsAutoencoders.fromFunction — Function

X = fromFunction(M::DensePolyManifold, fun)

Taylor expands Julia function fun to a polynomial, whose strcuture is prescribed by M.