Instantaneous frequencies, damping ratios

FoliationsManifoldsAutoencoders.ODEManifoldFrequencyDamping — Function

freq, damp, r = ODEManifoldFrequencyDamping(W, R, r0; output=nothing)

Calculates the instantaneous frequencies and damping ratios of the ROM $\boldsymbol{W}, \boldsymbol{R}$, where $\boldsymbol{R}$ is in the real normal form

\[\begin{pmatrix}\dot{z}_{1}\\ \dot{z}_{2} \end{pmatrix} = \begin{pmatrix}z_{1}f_{r}\left(z_{1}^{2}+z_{2}^{2}\right)-z_{2}f_{r}\left(z_{1}^{2}+z_{2}^{2}\right)\\ z_{1}f_{i}\left(z_{1}^{2}+z_{2}^{2}\right)+z_{2}f_{r}\left(z_{1}^{2}+z_{2}^{2}\right) \end{pmatrix}.\]

The input parameters are

W is $\boldsymbol{W}$: decoder or manifold immersion

S is $\boldsymbol{S}$: nonlinear map

r0: a vector of amplitudes that result is calculated at. This must be a reasonably fine mesh, because a finite difference is taken with respect to this mesh.

output: an optional linear map, that

FoliationsManifoldsAutoencoders.MAPManifoldFrequencyDamping — Function

freq, damp, r, freq_old, damp_old, r_old = MAPManifoldFrequencyDamping(W, S, r0, Tstep; output=nothing)

Calculates the instantaneous frequencies and damping ratios of the ROM $\boldsymbol{W}, \boldsymbol{S}$, where $\boldsymbol{S}$ is in the real normal form

\[\begin{pmatrix}z_{1}\\ z_{2} \end{pmatrix} \mapsto \begin{pmatrix}z_{1}f_{r}\left(z_{1}^{2}+z_{2}^{2}\right)-z_{2}f_{r}\left(z_{1}^{2}+z_{2}^{2}\right)\\ z_{1}f_{i}\left(z_{1}^{2}+z_{2}^{2}\right)+z_{2}f_{r}\left(z_{1}^{2}+z_{2}^{2}\right) \end{pmatrix}.\]

The input parameters are

W is $\boldsymbol{W}$: decoder or manifold immersion

S is $\boldsymbol{S}$: nonlinear map

Tstep is the time step that one application of $\boldsymbol{S}$ represents

r0: a vector of amplitudes that result is calculated at. This must be a reasonably fine mesh, because a finite difference is taken with respect to this mesh.

output: an optional linear map, that